// src/app/models/treeModels/strategicKpi.ts
export interface StrategicKpi {
    skpiId: number;
    name: string;
    quantify: string;
    remarks: string;
    year: number;
    groupId: number;
    visionId: number;
    plantId: number;
    isActive: boolean;
    isDeleted: boolean;
    createdOn: Date;
    createdBy: number;
    updatedOn: Date;
    updatedBy: number;
    deletedAt: Date;
    skpId: number;
}